clean: extract prepareArgument
